requirements.txt
aiogram
aiohttp
requests
python-dotenv
sqlalchemy